# Optimize performance of SecurityIdentifier.ToString()