in SqlClient when using trusted authentication the windows user identity is used as part of the pool key. Profiling has shown that getting the user sid is on the hot path for this and that optimizing the retrieval of the SecurityIdentifier string form will improve connection throughput. 

This change avoids the use of a StringBuilder, it's expansion and formatting intermediate strings by using a stackalloc span of the maximum possible length for a v1 string form sid and formatting directly into that buffer. The size calculation is included in the comments so the next person to look at this code doesn't have to work it out for themselves. 

The stackalloc buffer is 189 chars in length which looks odd but remember that chars are 16bit so this buffer is still 16bit aligned on the stack. It could be increased up to the closest power of 2 (192) if this is desirable, I know there are rules about this which @GrabYourPitchforks will need to review.

Performance changes make it almost 50% faster and change the memory allocation from using StringBuilder expansion and intermediate formatting representation strings to a flat allocation at the end of the method to return the string.

|      Method |     Mean |    Error |   StdDev | Ratio | Gen 0/1k Op |  Allocated Memory/Op |
------------ |---------:|---------:|--:|------:|-------:|--:|
   Reference | 829.0 ns | 9.954 ns | 8.824 ns |  1.00 |      0.4339 |  1368 B |
 Development | 431.9 ns | 3.390 ns | 3.171 ns |  0.52 |      0.1140 | 360 B |


/CC @GrabYourPitchforks , @bartonjs 
Contributes to https://github.com/dotnet/corefx/issues/30430 @saurabh500