@Wraith2 2018-11-18
The packaging build failure indicates that Span<> isn't found for the netcoreapp2.0 build I think. Any suggestions?