@jkotas 2018-11-18
> The packaging build failure indicates that Span<> isn't found for the netcoreapp2.0 build I think. Any suggestions?

Keep the existing code under ifdef for netstandard and netcoreapp2.0 configurations.

Alternatively, we may stop building netstandard and netcoreapp2.0 version of the package. @safern Is it an option for this package?