@safern 2018-11-20
You can just run dotnet msbuild /t:UpdateVSConfigurations in the root of your repo. Then just add the changes to the src\System.Security.Principal.Windows.csproj file to your commit.