@Wraith2 2018-11-21
```
D:\j\workspace\windows-TGrou---0d2c9ac4\src\Common\src\Interop\Windows\advapi32\Interop.LsaNtStatusToWinError.cs(12,9): error BCL0015: advapi32.dll!LsaNtStatusToWinError is not supported on one\more targeted platforms. [D:\j\workspace\windows-TGrou---0d2c9ac4\src\System.Security.Principal.Windows\src\System.Security.Principal.Windows.csproj]
```
Looks like it's coming from netcoreapp2.1-Windows_NT. That pinvoke is still in the `PinvokeAnalyzerExceptionList.analyzerdata.netcoreapp` file the same as it was before I started and which works for the 2.0 version of that target. Any guidance?