@safern 2018-11-26
You just need to duplicate this file: https://github.com/dotnet/corefx/blob/master/src/System.Security.Principal.Windows/src/PinvokeAnalyzerExceptionList.analyzerdata.netcoreapp and rename it to PinvokeAnalyzerExceptionList.analyzerdata.netcoreapp2.1